More fixes of unit tests for `RandUniIntRange`
Previously, we had unit tests that expected that every draw from different streams would result in a different random number.

This is does not correspond to the correct behavior of `RandUniIntRange`: instead, we can expect that at least some (usually a majority of) random numbers will be different among streams.