Enhancement for handling sparse missing weather input values
Previously, the weather generator could be used if weather input files were absent for entire years; otherwise, if a weather input file for a specific year was available, it was used. If that year contained a few missing values, they were impute with 0 for precipitation and with LOCF for temperature (i.e., last-observation-carried-forward).

Now, if the weather generator is turned on, it will be used to impute any day with missing values.

A new unit test covers this situation.