Attempt to fix `tools/run_doxygen.sh` for travis-ci
This works well locally, but travis-ci fails with
```
# Examine log file for remaining warnings/errors
warnings=$(grep "warning\|error" ${log})
grep "warning\|error" ${log}
makefile:260: recipe for target 'doc' failed
make: *** [doc] Error 1
The command "make doc" exited with 2.
```

It is not clear to me why `grep "warning\|error" ${log}` would exit with code 2. As an attempt, I catch grep's stderr and assign it to `warnings` as well.
It is also not clear whether or not travis-ci expects that this script (called by `make doc`) should define an exit value if we got warnings (see https://docs.travis-ci.com/user/job-lifecycle#how-does-this-work-or-why-you-should-not-use-exit-in-build-steps).