Memoization of `sun_angles` and `solar_radiation_extraterrestrial`
These functions return identical values for a doy for a given set of latitude, slope, and aspect that are fixed for a given SOILWAT2 run.
Instead of re-calculating `sun_angles` and `solar_radiation_extraterrestrial` for each year, we calculate the values once, store them in an array, and retrieve values during following simulation years.

- new function `SW_PET_init_run` to initialize memoized variables
- update extraterrestrial unit tests

Other changes:
- function `sun_hourangles` gained output argument `sun_angles`