Fix readability-redundant-casting
warning: redundant explicit casting to the same type 'char *' as the sub-expression, remove this casting [readability-redundant-casting]

`strrchr()` returns "char *" in C; however, it returns "const char *" if the first argument is "const char *" in C++.

`DirName()` and `BaseName()` call `strrchr()` with "const char *" as first argument.

Our variable `c` (that received the return value of `strrchr()`) was previously of type "char *"; thus, we needed to cast to "char *" to avoid loosing the "const" in C++; however, this was a redundant cast in C.
The solution was to define `c` with type "const char *": this is ok in C and in C++ because there is no need to cast.